make a whole bunch of files, import some data structures libraries
None of this is strictly functional right now (not working yet), but I'm hoping that once I clean it up it will be nicely functional. A lot of these files are coming from the imported datastructures library, but most of that libary (except for the license, readme, etc.) have been omitted.